--- a/packages/clerk-js/src/core/devBrowserHandler.ts
+++ b/packages/clerk-js/src/core/devBrowserHandler.ts
@@ -1,5 +1,4 @@
-import { buildURL, createCookieHandler, isDevOrStagingUrl, runIframe } from 'utils';
-
+import { buildURL, createCookieHandler, isDevOrStagingUrl, runIframe } from '../utils';
 import { DEV_BROWSER_SSO_JWT_HTTP_HEADER, DEV_BROWSER_SSO_JWT_KEY, DEV_BROWSER_SSO_JWT_PARAMETER } from './constants';
 import { clerkErrorDevInitFailed } from './errors';
 import { FapiClient } from './fapiClient';