--- a/packages/clerk-js/src/globals.d.ts
+++ b/packages/clerk-js/src/globals.d.ts
@@ -1,0 +1,5 @@
+declare global {
+  const __DEV__: boolean;
+  const __CLERKJS_VERSION__: string;
+}
+export {};