--- a/packages/clerk-js/src/ui/common/verification.ts
+++ b/packages/clerk-js/src/ui/common/verification.ts
@@ -1,5 +1,5 @@
 import { ClerkAPIError } from '@clerk/types';
-import { getClerkAPIErrorMessage, getGlobalError } from 'ui/common';
+import { getClerkAPIErrorMessage, getGlobalError } from 'ui/common/index';
 
 export const VerificationErrorMessage = {
   Incorrect: 'Incorrect, try again',