--- a/packages/clerk-js/src/ui/contexts/utils.ts
+++ b/packages/clerk-js/src/ui/contexts/utils.ts
@@ -1,5 +1,6 @@
 import { Clerk } from '@clerk/types';
-import { clerkCoreErrorContextProviderNotFound, clerkCoreErrorNoClerkSingleton } from 'core/errors';
+
+import { clerkCoreErrorContextProviderNotFound, clerkCoreErrorNoClerkSingleton } from '../../core/errors';
 
 export function assertClerkSingletonExists(clerk: Clerk | undefined): asserts clerk is Clerk {
   if (!clerk) {