--- a/packages/clerk-js/src/ui/router/PathRouter.tsx
+++ b/packages/clerk-js/src/ui/router/PathRouter.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { useCoreClerk } from 'ui/contexts';
-import { stripOrigin, trimTrailingSlash } from 'utils';
 
+import { stripOrigin, trimTrailingSlash } from '../../utils';
+import { useCoreClerk } from '../contexts';
 import { BaseRouter } from './BaseRouter';
 
 interface PathRouterProps {