--- a/packages/clerk-js/src/utils/memoizeStateListenerCallback.ts
+++ b/packages/clerk-js/src/utils/memoizeStateListenerCallback.ts
@@ -1,5 +1,6 @@
 import { ClientResource, ListenerCallback, Resources, SessionResource, UserResource } from '@clerk/types';
-import { Client, Session, User } from 'core/resources/internal';
+
+import { Client, Session, User } from '../core/resources/internal';
 
 type AcceptedResource = UserResource | ClientResource | SessionResource;