--- a/packages/clerk-js/src/v4/index.ts
+++ b/packages/clerk-js/src/v4/index.ts
@@ -1,0 +1,2 @@
+import Components from './Components';
+export default Components;