--- a/packages/clerk-js/src/v4/signIn/SignIn.tsx
+++ b/packages/clerk-js/src/v4/signIn/SignIn.tsx
@@ -1,0 +1,56 @@
+import { SignInProps } from '@clerk/types';
+import React from 'react';
+
+import { ComponentContext, useCoreClerk, useSignInContext, withCoreSessionSwitchGuard } from '../../ui/contexts';
+import { Route, Switch, VIRTUAL_ROUTER_BASE_PATH } from '../../ui/router';
+import { SignInStart } from './SignInStart';
+
+function RedirectToSignIn() {
+  const { redirectToSignIn } = useCoreClerk();
+  React.useEffect(() => {
+    void redirectToSignIn();
+  }, []);
+  return null;
+}
+
+function SignInRoutes(): JSX.Element {
+  const signInContext = useSignInContext();
+
+  return (
+    <Switch>
+      <Route index>
+        <SignInStart />
+      </Route>
+    </Switch>
+  );
+}
+
+SignInRoutes.displayName = 'SignIn';
+
+export const SignIn: React.ComponentType<SignInProps> = withCoreSessionSwitchGuard(SignInRoutes);
+
+export const SignInModal = (props: SignInProps): JSX.Element => {
+  const signInProps = {
+    signUpUrl: `/${VIRTUAL_ROUTER_BASE_PATH}/sign-up`,
+    ...props,
+  };
+
+  return (
+    <Route path='sign-in'>
+      <div className='cl-component cl-sign-in'>
+        <ComponentContext.Provider
+          value={{
+            componentName: 'SignIn',
+            ...signInProps,
+            routing: 'virtual',
+          }}
+        >
+          <SignIn
+            {...signInProps}
+            routing='virtual'
+          />
+        </ComponentContext.Provider>
+      </div>
+    </Route>
+  );
+};