--- a/packages/clerk-js/src/v4/signIn/index.ts
+++ b/packages/clerk-js/src/v4/signIn/index.ts
@@ -1,0 +1,1 @@
+export * from './SignIn';