--- a/tsconfig.eslint.json
+++ b/tsconfig.eslint.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "noEmit": true
   },
-  "include": ["packages/*/src"]
+  "include": ["packages/*/src", "packages/*/*.js"]
 }